Fix memory leak in house.cpp
There was a memory leak on HouseClass unit trackers because the
HouseClass destructor were not being called. Furthermore, it gets
very complicated to track all allocations of HouseClass in the code
to release them all properly.

This commit refactors the unit trackers and the houseclass so dynamic
memory allocation of the trackers are not necessary by passing the array
size as a template argument, resulting in no destructor requirement for
HouseClass, fixing the memory leak without ugly hacks.