Refactor a little of soundio_openal.cpp
The sound engine of this game is perhaps one of the most obscure pieces
of code in this repository. This commit refactors a little bit of it and
tries to somewhat explain how it works. It also removes some unecessary
memory allocations and stuff that was in the LockedData structure which
is not necessary.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>